(SERVER-2172) Create skeleton CLI w/ setup command
This implements a CLI class in addition to the Stub class currently
being used with the hope that we can buld a real featureful
implementation along side our existing stub.

The current CLI class does little, accepting --help flags with minimal
output, but includes tests.

There was discussion originally of both "import" and "setup" and
previously we used "import" in the stub impl., this moves us to use
"setup" in the non-stub implementation.