(SERVER-2172) Use inputs w/ SSL Store to validate leaf cert
The initial description of validation included:
  - all the certs in the bundle are valid authorities
  - any additional certs in the bundle are issuers of previous certs
    in the bundle
  - all of the crls, if provided, match the certs in the bundle
  - none of the certs in the bundle have been revoked when checking
    provided CRLs

And while writing this most of the meaning of criteria is met by
creating an SSL Store and validating the leaf cert with it. Specifically
a store validates that:
  - the leaf through root certs exist within the bundle
    (this does not validate that there are not other certs in the
    bundle, nor that they are in order save for the leaf cert being
    first).
  - none of the certs in the chain of trust have been revoked by their
    issuers, assuming CRLs are provided to check (there maybe revoked
    certs but they are not part of the path between leaf and root).
  - if provided, the certs in the chain of trust all have CRLs (but
    there may be additional CRLs provided).

The criteria expect the exact contents of the inputs to create a chain
of trust and no additional content, while this verifies the inputs at
a minimum create a valid chain of trust, but may have additional content.