(SERVER-2172) Provide constructors to create usable objects
Prior to this a `new` of either the PuppetConfig or X509Loader had some
values in a nil or partially initialized state. Where sensible we now
either do the loading/parsing in `initialize` or we provide a class
method that intializes and acts on the initialized class so the caller
is not expected to call multiple methods before being able to use the
object as desired.