(SERVER-2172) Add simple logger
Instead of directly accessing stdout/stderr we now provide a simple
logger to wrap those references. Changing verbosity is on the roadmap
and if this project is extended further pulling in the `logging` library
may be preferable. This is intended to provide an integration point for
those additional features in the future.