(SERVER-2172) Create serial and inventory if non-existant
Not originally part of this ticket, but a server will not start if some,
but not all, of its cert related files are present. This limits the
ability to actually use this cli.

Additional work is expected on this when we take on the full server
initialization work in SERVER-2175 & SERVER-2176.