(SERVER-2269) Allow specifying subject alt names when generating the CA
This commit adds the ability to specify subject alternative names for
the intermediate signing cert when using the `generate` action. These
can be configured via a flag, or read from the puppet.conf file. If none
are provided, Puppet's default set of alt names will be added. This
supports both DNS and IP alt names. If type is not specified, DNS names
are assumed.