(maint) Fix SSL Store creation
Prior to this we were creating an SSL Store for our HTTPS connection
using the improper API for adding CRLs, not configuring the CRL checking
flags correctly, passing in the agent's cert to add to the Store even
though it is unnecessary, and passing in the CA's cert bundle rather than
the agent's copy of it. (jeez)

This updates the creation of our SSL Store to be in line with how we
create them for agents (see puppet/ssl/host.rb in Puppet proper).