(SERVER-2304) Honor pre-existing master keys
The bootstrapping of Puppet Server's PKI very intentionally allows users
to have existing public and private keys for the master. Users must have
both or neither, only having one is an error. This is for some
bootstrapping of node workflows where an agent come pre-installed on an
image. See https://tickets.puppetlabs.com/browse/SERVER-528 for details
on the original rationale.

Previously, we did not make exceptions for the master's keys and
required that they not exist when running generate or import.

With this patch generate and import actions will now honor a
pre-existing keypair.