(SERVER-2305) Generate files needed for infra_crl
Previously we did not create the infra_inventory.txt, infra_serials, or
infra_crl.pem that is needed by the infra-crl feature of Puppet Sever.

This patch adds the creation of the first two files (as empty files) and
the last as a duplicate of the cacrl (they shouldn't diverge until nodes
have actually been revoked).

This has multiple caveats:
 - if someone in importing a "full" crl that will become the base of
   their "infra" crl
 - the locations of the files are hardcoded within specific file names
   to be rooted in $cadir (the current Puppet Server code allows them
   to changed, but we've discussed only allowing moving the cadir in
   Puppet Server).