(SERVER-2320) Allow generation of an authorized cert offline
The cert used by the CA CLI needs to have a special auth extension that
is disallowed by the Puppet Server CA, to prevent any node from
requesting a cert with the special permissions. Previously, this cert
was always created along with the rest of the CA files.

This commit adds a way to generate an authorized cert with the special
extension, from an existing CA. The special cert is signed offline, not
through the main Puppet CA. It can be generated with values for the
master, or with values for other nodes. In the latter case the cert and
keys will need to be distributed manually.

This also updates `generate` to not use the `subject_alt_names`
setting. This setting is supposed to be specific to the master node, but
`generate` can be used to create certs for any node, and those shouldn't
necessarily have the master's alt names. When using `generate`, the user
needs to specify both certname and any alt names manually.