(SERVER-2351) Check CA service before generating certs offline
In order to avoid race conditions with serial numbers, we want to
avoid allowing multiple processes to issue certificates at the same
time. This commit adds a check to make sure that Puppet Server's CA
service is offline before issuing certs out-of-band using Ruby code,
when using `generate --ca-client`. It queries the simple status endpoint
and errors if it receives a "running" response from the server.