(SERVER-2392) Respect server_list
Previously we did not look for the server_list value when computing the
value for the ca_server or ca_port. This makes us incompatible with
certain PE HA workflows. This was not an issue before because in our concrete
use cases the first value of server_list is $certname on the host in question.

However, in moving to use the same default for $server as the agent (which
improves our docker workflows) we broke our HA workflows. This updates us
to correctly respect the server_list (and prefer it over the $server setting.