(SERVER-2401) Update check_server_online to use new Exceptions.
Previously, we were catching ECONNREFUSED to determine if the server was
up. When we moved to custom error classes we no longer raise an
ECONNREFUSED, instead raising ConnectionFailed in it (and many other
potential exceptions) place.

So we should now be rescuing ConnectionFailed, but we only want to
swallow exceptions that originated from ECONNREFUSED, re-raising
ConnectionFaileds that originate from other exceptions.

To do so we add the ability to wrap an exception in our custom Error
class and update appropriate places where we were raising.