(SERVER-2559) Check CA for errors before enabling infra crl
This commit updates `puppetserver ca enable --infracrl` to check the CA
state for errors before attempting to generate the infrastructure CRL.
Prior to this commit, errors encountered when loading the CA state would
cause instance variables such as the certificate and private key to be
unset. This would result in CRL creation failing due to unexpected `nil`
values without providing clear messaging as to the root cause of the
problem.