(SERVER-2589) Add a --certname option to `list`
The new --certname option on the list action behaves similiarly to how
the option operates on other actions, in this case by limiting the
output to the certnames specified, and causing the command to return a
non-zero exit code in the event a certname specified is not able to be
listed (because it is not present).

This brings `puppetserver ca list` closer to parity with older commands
such as `puppet cert list`.

Initial implementation is not performance optimized. It doesn't look
like the certificate_statuses API actually supports returning a filtered
list, so this commit is forced to filter the results of requesting all
certificate statuses.