(SERVER-2552) Initialize leaf CRL during import
This commit updates the `puppetserver ca import` command to initialize
a CRL for the intermediate CA if the loader does not find one in the
`crl-chain` file. Not having a pre-generated CRL is the typical case as
creating a CRL requires access to the private key of the intermediate
CA. An external CA generally should not have access to this key and is
therefore not able to provide a pre-generated CRL along with the
intermediate CA cert.