(SERVER-2896) Update defaults for new cadir
With Puppet 7 the cadir location will move by default from `$ssldir` to
Puppet Server's confdir. However we want to respect any users that have
already created their cadir in either the old default location or a new
location. We also want to warn if they have it in the old location that
they should migrate the cadir. If users have a custom cadir but it
resides in the ssldir we also want to warn, a cadir within the ssldir
may cause accidental deletion of the cadir.