(SERVER-2895) Symlink to old cadir when setting up with the new default
This commit adds a method to symlink the new default cadir to the old default
cadir location when running the `setup` and `import` commands. This ensures
that, even on new installs, we can access the cadir from the old location
because there are still tools in the ecosystem that expect the old location.