(SERVER-2338) Ensure `ca-client` cert has certname as altname
We previously updated the certificate signing code in puppetserver
proper to always add the certname as a subject alt name when signing a
certificate. This commit updates the `generate` action of the CLI gem to
also ensure that the certname of the host the cert is being generated
for is added as an alt name, even if no other alt names are provided.
Because `generate` is used to create certs for other hosts, we can't use
the value of `certname` from settings in the CA code itself, but rather
need to use the value of `certname` that is passed into the command.