(SERVER-2740) Implement 'prune' action for Puppetserver's CA CLI
The purpose of this action is to deduplicate any revoked certificates the CRL issued by Puppet's CA cert.
The user may provides this action with a custom path to the 'puppet.conf' file through the use of the
'--config' option.  If the custom path is not provided, the program will use the default location of
the 'puppet.conf' file on disk.  Once the CRL is deduplicated, it will use the CA's private key that may
also be stored on disk to re-sign the CRL and update its version.  Using the '--verbose' option with this
subcommand will output on screen the information of the duplicated cert.