(SERVER-2842) Rearrange `--force` tests
This commit moves the tests for how `generate --ca-client` should behave when we
can't determine whether the puppetserver service is running to be next to the
test for how to behave when the puppetserver service *is* running. This commit
also improves the test descriptions to clarify the behavior being tested.