Adding a test case for failed lookups of CSRs
Note that this is just for the failure case that comes with #98's additions.
There's still a coverage gap that exists for the positive cases, but it was
like that when I got here and should probably be dealt with by someone who
knows ruby, rather than by my flailing here.