(SERVER-2797) Add failure exit code when `puppetserver ca list` hits errors.
When running a puppetserver ca command on a non-CA server, the command exits with a 0 = ok, when the command has obviously failed. The failure is easily detected by a human, but less so for a computer.

This seeks to distinguish "failed at looking up" from "found no certificates" and exit accordingly.