(PE-34952) Add inventory parsing module
This adds a module for parsing the inventory.txt file. This will handle any errors found in the file, and return a hash with certname as the key and the rest of the info has another hash.  Because the inventory will likely have old entries for a given certname, this records the serials for the old certnames, but the serial value will be the latest cert found.